refactor: rename file to index.ts
