+++ src/geometries/geometry.ts
@@ -5,7 +5,7 @@ import type { TileMatrixSet } from '../tiles/tileMatrixSet';
 import { positionToTileIndex, tileMatrixToBBox } from '../tiles/tiles';
 import type { CRS as CRSType, TileIndex } from '../tiles/types';
 import type { ArrayElement } from '../utils/types';
-import { validateCRS, validateCRSByOtherCRS, validateMetatile } from '../validations/validations';
+import { validateCRS, validateCRSByOtherCRS, validateMetatile } from '../validations';
 import type { CoordRefSysJSON, GeoJSONBaseGeometry, GeoJSONGeometry, JSONFGFeature } from './types';
 
 /**
