+++ src/geometries/point.ts
@@ -2,9 +2,9 @@ import { Tile } from '../tiles/tile';
 import type { TileMatrixSet } from '../tiles/tileMatrixSet';
 import { tileEffectiveHeight, tileEffectiveWidth, tileMatrixToBBox } from '../tiles/tiles';
 import type { TileMatrixId } from '../tiles/types';
-import { validateCRSByOtherCRS, validateMetatile, validatePointByTileMatrix, validateTileMatrixIdByTileMatrixSet } from '../validations/validations';
-import type { GeoJSONPoint, PointInput } from './types';
+import { validateCRSByOtherCRS, validateMetatile, validatePointByTileMatrix, validateTileMatrixIdByTileMatrixSet } from '../validations';
 import { BaseGeometry } from './baseGeometry';
+import type { GeoJSONPoint, PointInput } from './types';
 
 /**
  * Point geometry class
