+++ src/tiles/tile.ts
@@ -2,7 +2,7 @@ import { encodeToJSON } from '../crs/crs';
 import { BoundingBox } from '../geometries/boundingBox';
 import { Point } from '../geometries/point';
 import type { ArrayElement } from '../utils/types';
-import { validateMetatile, validateTileMatrix, validateTileMatrixIdByTileMatrixSet } from '../validations/validations';
+import { validateMetatile, validateTileMatrix, validateTileMatrixIdByTileMatrixSet } from '../validations';
 import type { TileMatrixSet } from './tileMatrixSet';
 import { TileRange } from './tileRange';
 import { tileEffectiveHeight, tileEffectiveWidth, tileMatrixToBBox } from './tiles';
